* Added support for Vista: HoboCopy will now prompt to run as an administrator * Modified logging - verbosity levels now more evenly divided. * /loglevel switch changed to /verbosity * Added support for copying only files that match specified patterns (e.g. *.txt) * Fixed bug where incremental copy was choking on access denied error * Default mode is now non-recursive: added /r switch to turn on recursive mode. * Added /simulate switch to simulate backup without snapshotting or copying. * Fixed major bug: components were not being selected properly for backup. As a result, state was crash-consistent only, even for writers that support the VSS API. * Fixed bug where delete of nonexistent directory would fail.

+++ hobocopy/trunk/CComException.h
@@ -55,7 +55,7 @@ public:
         // Hack: this part is not TCHAR-aware, but this is the only place
         size_t length = strlen(_file); 
         WCHAR* buffer = new WCHAR[length + 1]; 
-        ::MultiByteToWideChar(CP_ACP, 0, _file, length, buffer, length); 
+        ::MultiByteToWideChar(CP_ACP, 0, _file, (int) length, buffer, (int) length); 
         buffer[length] = L'\0';
         file.Empty(); 
         file.Append(buffer);
